gotur-app-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
SQLAlchemy>=2.0
pydantic[email]>=2.7
passlib[bcrypt]
PyJWT>=2.8
